feat(runtime): pass slot fallback into slot elements
Previously, we introduce if condition to the slot trait, which helps
render slot element conditionally. But it is not easy for a component
developer to implement fallback UI of a slot, because neither the
if condition is true or false, the slot elements function will return
the ImplWrapperMain component.
This patch bring a new way to implement fallback UI, by passing the
fallback UI as the second parameter of the slot elements function.