refactor(arco/Table): remove selectedRow state and add rowSelectChange events
seletedRow is a confusing property. It actually represents  the row you are currently working on.
In
the case of multiple selections, the selectedRow is the current row regardless of whether the
current row is selected or unselected. Although arco provides `selected` to determine if the current
row is selected, there is a lot of extra logic required to determine if the selectedRow has a
value.
And in single selection, `selectedRow` and `selectedRows` can be confusing